function_code/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-ai-vision-imageanalysis